+++ bitbots_ball_filter/setup.py
@@ -31,6 +31,7 @@ setup(
         'console_scripts': [
             'ball_filter = bitbots_ball_filter.ball_filter:main',
             'ball_publisher = bitbots_ball_filter.ball_publisher:main',
+            "ball_sim = bitbots_ball_filter.ball_sim:main",
         ],
     }
 )
